# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MINIMALL API
  description: |
    Ultra-fast link-in-bio platform API for Shopify merchants.
    
    ## Overview
    MINIMALL provides a comprehensive API for managing link-in-bio site configurations, 
    assets, analytics, and user interactions. The API is built with Next.js API routes
    and follows RESTful principles.
    
    ## Authentication
    Currently, the API uses shop-based authentication. Authentication improvements
    are planned for future releases.
    
    ## Rate Limiting
    API requests are subject to reasonable rate limits to ensure service stability.
    
    ## Error Handling
    All endpoints return consistent error responses with appropriate HTTP status codes
    and detailed error messages.
  version: 1.0.0
  contact:
    name: MINIMALL API Support
    email: api-support@minimall.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Admin API (Development)
  - url: http://localhost:3000
    description: Public API (Development)
  - url: https://admin.minimall.com
    description: Admin API (Production)
  - url: https://api.minimall.com
    description: Public API (Production)
paths:
...
  /api/configs/{configId}:
    get:
      tags:
        - Configurations
      summary: Get configuration
      description: Retrieve a site configuration by ID
      parameters:
        - name: configId
          in: path
          required: true
          description: Unique identifier for the site configuration
          schema:
            type: string
            example: "demo-shop-config"
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  config:
                    $ref: '#/components/schemas/SiteConfig'
                  source:
                    type: string
                    enum: [r2, database]
                    example: "r2"
        '404':
          description: Configuration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Configurations
      summary: Update configuration
      description: Update an existing site configuration
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteConfig'
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  config:
                    $ref: '#/components/schemas/SiteConfig'
                  message:
                    type: string
                    example: "Configuration updated successfully"
        '400':
          description: Invalid configuration data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Configurations
      summary: Delete configuration
      description: Delete a site configuration
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Configuration deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Configuration deleted successfully"
        '404':
          description: Configuration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    SiteConfig:
      type: object
      required:
        - id
        - shop
        - slug
        - settings
        - createdAt
      properties:
        id:
          type: string
          description: Unique identifier for the site configuration
          example: "demo-shop-config"
        shop:
          type: string
          description: Shopify shop domain
          example: "demo-shop.myshopify.com"
        slug:
          type: string
          description: URL slug for the site
          example: "demo-shop"
        content:
          type: array
          description: Array of content items for the site
          items:
            $ref: '#/components/schemas/ContentItem'
        settings:
          type: object
          required:
            - shopDomain
          properties:
            shopDomain:
              type: string
              example: "demo-shop.myshopify.com"
            theme:
              type: object
              properties:
                primaryColor:
                  type: string
                  example: "#2563eb"
                backgroundColor:
                  type: string
                  example: "#ffffff"
                textColor:
                  type: string
                  example: "#1f2937"
            seo:
              type: object
              properties:
                title:
                  type: string
                  example: "Demo Shop - Link in Bio"
                description:
                  type: string
                  example: "Check out our amazing products!"
                keywords:
                  type: string
                  example: "shop, products, fashion"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00Z"
    ContentItem:
      type: object
      required:
        - id
        - type
        - position
        - isVisible
      properties:
        id:
          type: string
          example: "image-123"
        type:
          type: string
          enum: [image, video, product, text, link, social]
          example: "image"
        title:
          type: string
          example: "Hero Image"
        position:
          type: integer
          minimum: 1
          example: 1
        isVisible:
          type: boolean
          example: true
        description:
          type: string
          example: "Main promotional image"
      discriminator:
        propertyName: type
        mapping:
          image: '#/components/schemas/ImageContent'
          video: '#/components/schemas/VideoContent'
          product: '#/components/schemas/ProductContent'
          text: '#/components/schemas/TextContent'
          link: '#/components/schemas/LinkContent'
          social: '#/components/schemas/SocialContent'
    ImageContent:
      allOf:
        - $ref: '#/components/schemas/ContentItem'
        - type: object
          properties:
            src:
              type: string
              format: uri
              example: "https://cdn.shopify.com/s/files/1/example/image.jpg"
            alt:
              type: string
              example: "Product showcase image"
            width:
              type: integer
              example: 400
            height:
              type: integer
              example: 300
    VideoContent:
      allOf:
        - $ref: '#/components/schemas/ContentItem'
        - type: object
          properties:
            src:
              type: string
              format: uri
              example: "https://example.com/video.mp4"
            thumbnail:
              type: string
              format: uri
              example: "https://example.com/thumb.jpg"
            autoplay:
              type: boolean
              example: false
    ProductContent:
      allOf:
        - $ref: '#/components/schemas/ContentItem'
        - type: object
          properties:
            productId:
              type: string
              example: "gid://shopify/Product/123456789"
            handle:
              type: string
              example: "sample-product"
            showPrice:
              type: boolean
              example: true
            showDescription:
              type: boolean
              example: true
    TextContent:
      allOf:
        - $ref: '#/components/schemas/ContentItem'
        - type: object
          properties:
            content:
              type: string
              example: "Welcome to our store!"
            style:
              type: object
              properties:
                fontSize:
                  type: integer
                  example: 16
                color:
                  type: string
                  example: "#000000"
                textAlign:
                  type: string
                  enum: [left, center, right]
                  example: "center"
    LinkContent:
      allOf:
        - $ref: '#/components/schemas/ContentItem'
        - type: object
          properties:
            href:
              type: string
              format: uri
              example: "https://shop.example.com"
            target:
              type: string
              enum: [_blank, _self]
              example: "_blank"
            style:
              type: object
              properties:
                buttonStyle:
                  type: string
                  enum: [primary, secondary, outline]
                  example: "primary"
    SocialContent:
      allOf:
        - $ref: '#/components/schemas/ContentItem'
        - type: object
          properties:
            platform:
              type: string
              enum: [instagram, tiktok, twitter, facebook, youtube]
              example: "instagram"
            username:
              type: string
              example: "@shopexample"
            showFollowers:
              type: boolean
              example: true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Configuration not found"
    ValidationError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid configuration data"
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
                example: "invalid_type"
              expected:
                type: string
                example: "string"
              received:
                type: string
                example: "number"
              path:
                type: array
                items:
                  type: string
                example: ["settings", "shopDomain"]
              message:
                type: string
                example: "Expected string, received number"
...